attempt to upgrade to react 0.13.
