+++ package.json
@@ -16,7 +16,6 @@
     "gulp-less": "^3.0.1",
     "gulp-minify-css": "^1.0.0",
     "gulp-plumber": "^0.6.6",
-    "gulp-html-prettify": "^0.0.1",
     "gulp-rename": "^1.2.0",
     "gulp-s3": "^0.3.0",
     "gulp-sitemap": "^2.0.1",
@@ -35,12 +34,11 @@
     "mofo-style": "^1.0.1",
     "node-jsx": "^0.12.4",
     "phantomjs": "^1.9.16",
-    "react": "^0.12.2",
-    "react-a11y": "0.0.6",
-    "react-bootstrap": "^0.13.1",
-    "react-ga": "^1.0.8",
-    "react-router": "^0.12.4",
-    "react-select": "^0.4.2",
+    "react": "^0.13.1",
+    "react-bootstrap": "^0.20.3",
+    "react-ga": "^1.0.12",
+    "react-router": "^0.13.2",
+    "react-select": "^0.4.6",
     "should": "^5.1.0",
     "superagent": "^1.1.0",
     "underscore": "^1.8.2",
